HolidayBooking testing
Created a debug flow from 'My Holidays' - by clicking amend booking, the user is shown a 'debug page' - the HolidayBooking object is deconstructed using its internal deconstruct method, then Reconstructed using its 'Reconstruct' method. The class was designed this way so there was no need for serialisation, and no need to inherit something like Parcelable.

Object is deconstructed into a Bundle on the 'My Holidays' page, then reconstructed from the Bundle to the object on the 'Amend Booking' page. The Fragment's Java class then populates the layout with the booking information. This was intended as a method of debugging / testing the effectiveness of object traversal using 'home-grown' methods.